03_my_implementations/chapter4_serving_patterns/02_synchronous_pattern/client/requirements.txt
tensorflow-serving-api>=2.15.0
grpcio>=1.60.0
numpy>=1.24.0
requests>=2.31.0
